Ripped apart the SensorTagLib's DeviceSelectActivity to get its SensorTag/Bluetooth connectivity functionality without any GUI stuff. Created the SensorTagConnector class which handles connections to the SensorTag. Successfully connected to the SensorTag and obtained readings through the HHM and Listener.

+++ app/src/main/java/me/dotteam/dotprod/SensorListenerInterface.java
@@ -10,6 +10,6 @@ public interface SensorListenerInterface {
         TEMPERATURE, HUMIDITY, PRESSURE, PEDOMETER, MAGNETOMETER
     }
 
-    void update(HikeSensors hikesensors, float value);
+    void update(HikeSensors hikesensors, double value);
 
 }
